Dynamic threading is working.
