+++ engine/src/flutter/lib/web_ui/skwasm/library_skwasm_support.js
@@ -8,11 +8,51 @@
 mergeInto(LibraryManager.library, {
   $skwasm_support_setup__postset: 'skwasm_support_setup();',
   $skwasm_support_setup: function() {
-    console.log("skwasmSingleThreaded: " + Module["skwasmSingleThreaded"]);
+    // This value represents the difference between the time origin of the main
+    // thread and whichever web worker this code is running on. This is so that
+    // when we report frame timings, that they are in the same time domain
+    // regardless of whether they are captured on the main thread or the web
+    // worker.
+    let timeOriginDelta = 0;
+    skwasm_registerMessageListener = function(threadId, listener) {
+      const eventListener = function({data}) {
+        const skwasmMessage = data.skwasmMessage;
+        if (!skwasmMessage) {
+          return;
+        }
+        if (skwasmMessage == 'syncTimeOrigin') {
+          timeOriginDelta = performance.timeOrigin - data.timeOrigin;
+          return;
+        }
+        listener(data);
+      };
+      if (!threadId) {
+        addEventListener("message", eventListener);
+      } else {
+        _wasmWorkers[threadId].addEventListener("message", eventListener);
+        _wasmWorkers[threadId].postMessage({
+          skwasmMessage: 'syncTimeOrigin',
+          timeOrigin: performance.timeOrigin,
+        });
+      }
+    };
+    skwasm_getCurrentTimestamp = function() {
+      return performance.now() + timeOriginDelta;
+    };
+    skwasm_postMessage = function(message, transfers, threadId) {
+      if (threadId) {
+        _wasmWorkers[threadId].postMessage(message, transfers);
+      } else {
+        postMessage(message, transfers);
+      }
+    };
 
     const handleToCanvasMap = new Map();
     const associatedObjectsMap = new Map();
 
+    _skwasm_isSingleThreaded = function() {
+      return Module["skwasmSingleThreaded"];
+    };
     _skwasm_setAssociatedObjectOnThread = function(threadId, pointer, object) {
       skwasm_postMessage({
         skwasmMessage: 'setAssociatedObject',
@@ -177,15 +217,22 @@ mergeInto(LibraryManager.library, {
       });
     }
   },
-  $skwasm_support_setup__deps: [ '$skwasm_threading_setup'],
+  $skwasm_registerMessageListener: function() {},
+  $skwasm_registerMessageListener__deps: ['$skwasm_support_setup'],
+  $skwasm_getCurrentTimestamp: function () {},
+  $skwasm_getCurrentTimestamp__deps: ['$skwasm_support_setup'],
+  $skwasm_postMessage: function () {},
+  $skwasm_postMessage__deps: ['$skwasm_support_setup'],
+  skwasm_isSingleThreaded: function() {},
+  skwasm_isSingleThreaded__deps: ['$skwasm_support_setup'],
   skwasm_setAssociatedObjectOnThread: function () {},
-  skwasm_setAssociatedObjectOnThread__deps: ['$skwasm_support_setup'],
+  skwasm_setAssociatedObjectOnThread__deps: ['$skwasm_support_setup', '$skwasm_postMessage'],
   skwasm_getAssociatedObject: function () {},
   skwasm_getAssociatedObject__deps: ['$skwasm_support_setup'],
   skwasm_disposeAssociatedObjectOnThread: function () {},
   skwasm_disposeAssociatedObjectOnThread__deps: ['$skwasm_support_setup'],
   skwasm_connectThread: function() {},
-  skwasm_connectThread__deps: ['$skwasm_support_setup'],
+  skwasm_connectThread__deps: ['$skwasm_support_setup', '$skwasm_registerMessageListener', '$skwasm_getCurrentTimestamp'],
   skwasm_dispatchRenderPictures: function() {},
   skwasm_dispatchRenderPictures__deps: ['$skwasm_support_setup'],
   skwasm_createOffscreenCanvas: function () {},
