+++ engine/src/flutter/lib/web_ui/skwasm/skwasm_support.h
@@ -13,6 +13,7 @@
 using SkwasmObject = __externref_t;
 
 extern "C" {
+extern bool skwasm_isSingleThreaded();
 extern void skwasm_setAssociatedObjectOnThread(unsigned long threadId,
                                                void* pointer,
                                                SkwasmObject object);
