+++ engine/src/flutter/web_sdk/BUILD.gn
@@ -414,7 +414,6 @@ if (!is_fuchsia) {
         "//flutter/third_party/canvaskit:canvaskit_chromium_group",
         "//flutter/third_party/canvaskit:canvaskit_group",
         "//flutter/third_party/canvaskit:skwasm_group",
-        "//flutter/third_party/canvaskit:skwasm_st_group",
       ]
     }
     deps += [ "//flutter/lib/web_ui/flutter_js" ]
@@ -440,9 +439,6 @@ if (!is_fuchsia) {
         "$root_out_dir/flutter_web_sdk/canvaskit/skwasm.js",
         "$root_out_dir/flutter_web_sdk/canvaskit/skwasm.js.symbols",
         "$root_out_dir/flutter_web_sdk/canvaskit/skwasm.wasm",
-        "$root_out_dir/flutter_web_sdk/canvaskit/skwasm_st.js",
-        "$root_out_dir/flutter_web_sdk/canvaskit/skwasm_st.js.symbols",
-        "$root_out_dir/flutter_web_sdk/canvaskit/skwasm_st.wasm",
       ]
     }
